+++ aten/src/ATen/native/vulkan/api/Shader.cpp
@@ -1,9 +1,5 @@
 #include <ATen/native/vulkan/api/Shader.h>
 
-#ifdef USE_VULKAN_SHADERC_RUNTIME
-#include <shaderc/shaderc.hpp>
-#endif /* USE_VULKAN_SHADERC_RUNTIME */
-
 namespace at {
 namespace native {
 namespace vulkan {
@@ -14,38 +10,19 @@ namespace api {
 //
 
 ShaderInfo::ShaderInfo()
-    : type(ShaderInfo::Type::SPIRV),
-      src_code{
-          .spirv =
-              {
-                  nullptr,
-                  0u,
-              },
+    : src_code{
+          nullptr,
+          0u,
       } {}
-
-ShaderInfo::ShaderInfo(std::string name, const char* const glsl_src)
-    : type(ShaderInfo::Type::GLSL),
-      src_code{
-          .glsl =
-              {
-                  glsl_src,
-                  0u,
-              },
-      },
-      kernel_name{std::move(name)} {}
 
 ShaderInfo::ShaderInfo(
     std::string name,
     const uint32_t* const spirv_bin,
     const uint32_t size,
     const std::vector<VkDescriptorType>& layout)
-    : type(Type::SPIRV),
-      src_code{
-          .spirv =
-              {
-                  spirv_bin,
-                  size,
-              },
+    : src_code{
+          spirv_bin,
+          size,
       },
       kernel_name{std::move(name)},
       kernel_layout{layout} {}
@@ -58,13 +35,9 @@ ShaderInfo::ShaderInfo(
     const std::vector<uint32_t>& tile_size,
     const StorageType bias_storage_type,
     const StorageType weight_storage_type)
-    : type(Type::SPIRV),
-      src_code{
-          .spirv =
-              {
-                  spirv_bin,
-                  size,
-              },
+    : src_code{
+          spirv_bin,
+          size,
       },
       kernel_name{std::move(name)},
       kernel_layout{layout},
@@ -77,17 +50,9 @@ ShaderInfo::ShaderInfo(
 }
 
 bool operator==(const ShaderInfo& _1, const ShaderInfo& _2) {
-  if (_1.type != _2.type) {
-    return false;
-  }
-
-  if (_1.type == ShaderInfo::Type::SPIRV) {
-    return (
-        _1.src_code.spirv.bin == _2.src_code.spirv.bin &&
-        _1.src_code.spirv.size == _2.src_code.spirv.size);
-  } else {
-    return (_1.src_code.glsl.src == _2.src_code.glsl.src);
-  }
+  return (
+      _1.src_code.bin == _2.src_code.bin &&
+      _1.src_code.size == _2.src_code.size);
 }
 
 //
@@ -153,8 +118,8 @@ void swap(ShaderLayout& lhs, ShaderLayout& rhs) noexcept {
 
 ShaderModule::ShaderModule(const VkDevice device, const ShaderInfo& source)
     : device_(device), handle_{VK_NULL_HANDLE} {
-  const uint32_t* code = source.src_code.spirv.bin;
-  uint32_t size = source.src_code.spirv.size;
+  const uint32_t* code = source.src_code.bin;
+  uint32_t size = source.src_code.size;
 
   const VkShaderModuleCreateInfo shader_module_create_info{
       VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO, // sType
