+++ scripts/build_android.sh
@@ -157,9 +157,6 @@ if [ -n "${USE_VULKAN}" ]; then
   if [ -n "${USE_VULKAN_RELAXED_PRECISION}" ]; then
     CMAKE_ARGS+=("-DUSE_VULKAN_RELAXED_PRECISION=ON")
   fi
-  if [ -n "${USE_VULKAN_SHADERC_RUNTIME}" ]; then
-    CMAKE_ARGS+=("-DUSE_VULKAN_SHADERC_RUNTIME=ON")
-  fi
 fi
 
 # Use-specified CMake arguments go last to allow overridding defaults
